+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/repository/ObjectMetadataRepository.java
@@ -1,0 +1,10 @@
+package org.ubiquia.core.flow.repository;
+
+
+import org.ubiquia.common.library.api.repository.AbstractEntityRepository;
+import org.ubiquia.common.model.ubiquia.entity.ObjectMetadataEntity;
+
+public interface ObjectMetadataRepository
+    extends AbstractEntityRepository<ObjectMetadataEntity> {
+
+}