+++ common/java/library/api/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] 2026-03-30
+### Added
+- `findEntitiesNeedingSync()` query to `AbstractEntityRepository` to identify records with no sync history or with updates newer than their last sync
+
 ## [0.1.0] 2025-06-02
 ### Added
 - Initial commit