+++ common/java/model/ubiquia/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] 2026-03-30
+### Added
+- `SyncEntity` to track which agents have received a given model record
+- `Sync` DTO counterpart for `SyncEntity`
+
 ## [0.16.1] 2026-03-03
 ### Fixed
 - Null pointer exception. Clearly Java's fault, and not the author.